+++ views/update/update-manage-trusts-individuals-or-entities-involved.html
@@ -163,8 +163,7 @@
                 },
                 actions: {
                   items: [ CREATE_CHANGE_LINK(
-                    OE_CONFIGS.UPDATE_MANAGE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL + "/" + pageData.trustData.trustId
-                    + OE_CONFIGS.TRUST_LEGAL_ENTITY_BENEFICIAL_OWNER_URL + "/" + trustee.id + OE_CONFIGS.LEGAL_ENTITY_NAME,
+                    OE_CONFIGS.UPDATE_MANAGE_TRUSTS_TELL_US_ABOUT_THE_LEGAL_ENTITY_URL + "/" + trustee.id + OE_CONFIGS.LEGAL_ENTITY_NAME,
                     "Legal Entity Trustee " + trustee.name,
                     "change-legal-entity-trustee-button"
                 ) ]
@@ -183,8 +182,7 @@
                 },
                 actions: {
                   items: [ CREATE_CHANGE_LINK(
-                    OE_CONFIGS.UPDATE_MANAGE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL + "/" + pageData.trustData.trustId
-                    + OE_CONFIGS.TRUST_LEGAL_ENTITY_BENEFICIAL_OWNER_URL + "/" + trustee.id + OE_CONFIGS.LEGAL_ENTITY_NAME,
+                    OE_CONFIGS.UPDATE_MANAGE_TRUSTS_TELL_US_ABOUT_THE_LEGAL_ENTITY_URL + "/" + trustee.id + OE_CONFIGS.LEGAL_ENTITY_NAME,
                     "Legal Entity Trustee " + trustee.name,
                     "change-legal-entity-trustee-button"
                 ) ]
